Update sw_setup.rst
`make modules_install` fails due to permission issues in the unpacked rootfs.  

If run with sudo, the current kernel config is ignored and a prompt is entered.  Additional steps have been added to change the lib/modules and lib/firmware directories to be owned by the current user prior to the module_install command, then ownership is reverted back to the root user.